form hooks simplified
