+++ src/lib/hooks/useJoinUsFrom.tsx
@@ -1,68 +1,37 @@
-import { useEffect, useState } from 'react'
-import {
-  JoinUsFieldsValidity,
-  JoinUsFormFields,
-  validateField,
-} from '~/utils/form-validation'
+import { JoinUsFieldsValidity, JoinUsFormFields } from '~/utils/form-validation'
+import { useForm } from '~/lib/hooks/useFrom'
 
 export const useJoinUsFrom = (
   formInitState: JoinUsFormFields,
   validityInitState: JoinUsFieldsValidity,
 ) => {
-  const [validForm, setValidForm] = useState<boolean>(false)
-  const [formData, setFormData] = useState<JoinUsFormFields>(formInitState)
-  const [validFields, setValidFields] =
-    useState<JoinUsFieldsValidity>(validityInitState)
-  const [isSuccess, setSuccess] = useState<boolean>(true)
-  const [isSubmitted, setSubmitted] = useState<boolean>(false)
-  const [isLoading, setIsLoading] = useState(false)
-
-  useEffect(() => {
-    setValidForm(Object.values(validFields).every(Boolean))
-  }, [validFields])
-
-  useEffect(() => {
-    if (isSubmitted) {
-      setTimeout(() => {
-        setSubmitted(false)
-      }, 5000)
-    }
-  }, [isSubmitted])
-
-  const updateFormField = ({ id, value }) => {
-    setFormData({
-      ...formData,
-      [id]: value,
-    })
-    setValidFields((current) => ({
-      ...current,
-      [id]: validateField(id, value),
-    }))
-  }
-
-  const handleSubmit = async (submit: () => void) => {
-    try {
-      setIsLoading(true)
-      submit()
-      setSuccess(true)
-      setFormData(formInitState)
-      setValidFields(validityInitState)
-    } catch (error) {
-      setSuccess(false)
-    } finally {
-      setSubmitted(true)
-      setIsLoading(false)
-    }
-  }
-
-  return {
-    validForm,
+  const {
     formData,
     validFields,
     isSuccess,
     isSubmitted,
     isLoading,
-    handleSubmit,
+    validForm,
+    handleNext,
     updateFormField,
+    handleSubmit,
+    saveFormData,
+    setAllFormField,
+    validateFields,
+  } = useForm(formInitState, validityInitState, '')
+
+  return {
+    formData,
+    validFields,
+    isSuccess,
+    isSubmitted,
+    isLoading,
+    validForm,
+    handleNext,
+    updateFormField,
+    handleSubmit,
+    saveFormData,
+    setAllFormField,
+    validateFields,
   }
 }
